Reduce folder nesting, update workflows to match
