+++ .github/workflows/dotnet-core-desktop.yml
@@ -76,4 +76,4 @@ jobs:
 
     # Execute all unit tests in the solution
     - name: Execute unit tests
-      run: dotnet test Cosmic.NET/Cosmic.NET.sln
+      run: dotnet test Cosmic.NET.sln
