+++ Cosmic.NET.WPF/App.config
