+++ Cosmic.NET.WPF/App.xaml
