+++ Cosmic.NET.WPF/App.xaml.cs
