+++ Cosmic.NET.WPF/Cosmic.NET.WPF.csproj
