+++ Cosmic.NET.WPF/MainWindow.xaml.cs
