+++ Cosmic.NET.WPF/MainWindowVM.cs
