+++ Cosmic.NET.WPF/Resources/CloseWindow.png
