+++ Cosmic.NET.WPF/Resources/CloseWindowWhite.png
