+++ Cosmic.NET.WPF/Resources/wedge.png
