+++ Cosmic.NET.sln
@@ -1,0 +1,34 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cosmic.NET", "Cosmic.NET\Cosmic.NET.csproj", "{C862934A-F83A-4BF6-8D59-E952DE5A4968}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cosmo", "cosmo\Cosmo.csproj", "{6832DDEB-C0F3-4AF2-846C-FB8ED3494F42}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cosmic.NET.WPF", "Cosmic.NET.WPF\Cosmic.NET.WPF.csproj", "{1EE3F27D-7A23-41FA-9458-C8E750F6E7E2}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C862934A-F83A-4BF6-8D59-E952DE5A4968}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C862934A-F83A-4BF6-8D59-E952DE5A4968}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C862934A-F83A-4BF6-8D59-E952DE5A4968}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C862934A-F83A-4BF6-8D59-E952DE5A4968}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6832DDEB-C0F3-4AF2-846C-FB8ED3494F42}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6832DDEB-C0F3-4AF2-846C-FB8ED3494F42}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6832DDEB-C0F3-4AF2-846C-FB8ED3494F42}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6832DDEB-C0F3-4AF2-846C-FB8ED3494F42}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1EE3F27D-7A23-41FA-9458-C8E750F6E7E2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1EE3F27D-7A23-41FA-9458-C8E750F6E7E2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1EE3F27D-7A23-41FA-9458-C8E750F6E7E2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1EE3F27D-7A23-41FA-9458-C8E750F6E7E2}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
