+++ Cosmic.NET.sln.DotSettings
