--- Cosmic.NET/Cosmic.NET/Cosmic.NET.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>net6.0-windows</TargetFramework>
-    <UseWindowsForms>true</UseWindowsForms>
-    <Version>1.2.1</Version>
-    <AssemblyVersion>1.2.1</AssemblyVersion>
-    <FileVersion>1.2.1</FileVersion>
-    <Copyright>Copyright © Joshua Kempner 2008-2022</Copyright>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\cosmo\Cosmo.csproj" />
-  </ItemGroup>
-</Project>
