+++ cosmo/Cosmo.csproj
@@ -1,0 +1,12 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <Version>1.2.1</Version>
+    <AssemblyVersion>1.2.1</AssemblyVersion>
+    <FileVersion>1.2.1</FileVersion>
+    <Copyright>Copyright © Joshua Kempner 2008-2022</Copyright>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="reactiveui" Version="18.3.1" />
+  </ItemGroup>
+</Project>
